- move route constants to constants file - allow to click next with optional framework - WIP: allow next on PROJECT TYPE page - revert project Type to FULL STACK so extension works
