+++ src/client/src/components/Footer/Footer.test.tsx
@@ -6,7 +6,7 @@ import { RenderResult } from "@testing-library/react";
 import { renderWithStore } from "../../testUtils";
 import styles from "./styles.module.css";
 import buttonStyles from "../../css/buttonStyles.module.css";
-import { ROUTE } from "../../utils/routes/constants";
+import { ROUTE } from "../../utils/constants/constants";
 import { AppState } from "../../store/combineReducers";
 
 describe("Footer", () => {
