+++ src/client/src/pages/PageAddServices/ServiceCard/index.tsx
@@ -8,8 +8,7 @@ import { isLoggedInSelector } from "../../../store/config/azure/selector";
 import { setDetailPageAction } from "../../../store/config/detailsPage/action";
 import { AppState } from "../../../store/combineReducers";
 
-import { KEY_EVENTS } from "../../../utils/constants/constants";
-import { ROUTE } from "../../../utils/routes/constants";
+import { KEY_EVENTS, ROUTE } from "../../../utils/constants/constants";
 
 import Icon from "../../../components/Icon";
