+++ src/client/src/store/userSelection/app/wizardSelectionSelector/wizardSelectionSelector.test.ts
@@ -1,6 +1,6 @@
 import configureMockStore from "redux-mock-store";
 import { isEnableNextPageSelector} from "./wizardSelectionSelector";
-import { ROUTE } from "../../../../utils/routes/constants";
+import { ROUTE } from "../../../../utils/constants/constants";
 import { getInitialState } from "../../../../mockData/mockStore";
 import { AppState } from "../../../combineReducers";
 import { ISelected } from "../../../../types/selected";
