--- src/client/src/utils/routes/constants.ts
@@ -1,13 +1,0 @@
-const ROUTE = {
-  PAGE_DETAILS: "/PageDetail",
-  SELECT_FRAMEWORKS: "/SelectFrameworks",
-  SELECT_PROJECT_TYPE: "/SelectProjectType",
-  ADD_PAGES: "/AddPages",
-  ADD_SERVICES: "/AddServices",
-  REVIEW_AND_GENERATE: "/ReviewAndGenerate",
-  NEW_PROJECT: "/",
-};
-
-
-
-export { ROUTE };
